--- client/src/components/layout/VersionDisplay/index.js
@@ -1,1 +1,0 @@
-export { default } from './VersionDisplay';
