--- client/src/components/layout/VersionProvider/index.js
@@ -1,1 +1,0 @@
-export { default } from './VersionProvider';
