+++ client/src/components/layout/components/NavItems.jsx
@@ -1,6 +1,6 @@
 import NavItem from './NavItem';
 
-import navItemList from './navItems.json';
+import navItemList from '../navItems.json';
 
 export default function NavItems() {
   return navItemList.map((navItem) => (
